+++ src/app/app-routing.ts
@@ -1,6 +1,7 @@
 import { Routes } from '@angular/router';
 import { AutoLoginPartialRoutesGuard } from 'angular-auth-oidc-client';
-import {basicGuard, settingsGuard} from './core/guards/accessLevel.guard'
+import {basicGuard, settingsGuard} from './core/guards/accessLevel.guard';
+
 export const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: 'home' },
   {
@@ -14,27 +15,25 @@ export const routes: Routes = [
   },
   {
     path: 'organization/credentials',
-    loadChildren: () =>
-      import(
-        './features/credential-management/credential-management.routes'
-        ).then((m) => m.default),
     canActivate: [AutoLoginPartialRoutesGuard, basicGuard],
-  },
-  {
-    path: 'organization/credentials/create',
-    loadChildren: () =>
-      import('./features/credential-issuance/credential-issuance.routes').then(
-        (m) => m.default
-      ),
-    canActivate: [AutoLoginPartialRoutesGuard, basicGuard],
-  },
-  {
-    path: 'organization/credentials/create2/:id',
-    loadChildren: () =>
-      import('./features/credential-issuance/credential-issuance.routes').then(
-        (m) => m.default
-      ),
-    canActivate: [AutoLoginPartialRoutesGuard, basicGuard],
+    children: [
+      {
+        path: '',
+        loadChildren: () => import('./features/credential-management/credential-management.routes').then(m => m.default),
+      },
+      {
+        path: 'details',
+        loadChildren: () => import('./features/credential-details/credential-details.routes').then(m => m.default),
+      },
+      {
+        path: 'create',
+        loadChildren: () => import('./features/credential-issuance/credential-issuance.routes').then(m => m.default),
+      },
+      {
+        path: 'create-as-signer',
+        loadChildren: () => import('./features/credential-issuance/credential-issuance.routes').then(m => m.default),
+      },
+    ],
   },
   {
     path: 'credential-offer',