+++ src/app/core/services/credential-procedure.service.spec.ts
@@ -3,14 +3,16 @@ import {HttpTestingController, provideHttpClientTesting} from '@angular/common/h
 import {CredentialProcedureService} from './credential-procedure.service';
 import {environment} from 'src/environments/environment';
 import {HttpErrorResponse, provideHttpClient} from '@angular/common/http';
-import {ProcedureRequest} from '../models/dto/procedure-request.dto';
+
 import {ProcedureResponse} from "../models/dto/procedure-response.dto";
-import {LearCredentialEmployeeDataDetail} from "../models/dto/lear-credential-employee-data-detail.dto";
+
 import {throwError} from 'rxjs';
 import {DialogWrapperService} from "../../shared/components/dialog/dialog-wrapper/dialog-wrapper.service";
 import {TranslateService} from "@ngx-translate/core";
 import {Router} from "@angular/router";
 import {API} from "../constants/api.constants";
+import { LEARCredentialDataDetails } from '../models/entity/lear-credential';
+import { EmployeeProcedureRequest } from '../models/dto/procedure-request.dto';
 
 const notFoundErrorResp = new HttpErrorResponse({
   error: '404 error',
@@ -95,8 +97,8 @@ describe('CredentialProcedureService', () => {
 
   it('should fetch credential procedure by id successfully', () => {
     const procedureId = '1';
-    const mockData: LearCredentialEmployeeDataDetail =
-      { procedure_id: '1', credential_status: 'completed', credential: { mandatee: {}, mandator: {}, power: [] } as any }
+    const mockData: LEARCredentialDataDetails =
+      { procedure_id: '1', credential_status: 'VALID', credential: { mandatee: {}, mandator: {}, power: [] } as any }
     ;
 
     service.getCredentialProcedureById(procedureId).subscribe(data => {
@@ -123,7 +125,7 @@ describe('CredentialProcedureService', () => {
   });
 
   it('should save credential procedure successfully', () => {
-    const IssuanceRequestMock:ProcedureRequest = {
+    const IssuanceRequestMock:EmployeeProcedureRequest = {
       schema: "LEARCredentialEmployee",
       format: "jwt_vc_json",
       payload: {
@@ -153,7 +155,7 @@ describe('CredentialProcedureService', () => {
   });
 
   it('should handle error when saving credential procedure', () => {
-    const IssuanceRequestMock:ProcedureRequest = {
+    const IssuanceRequestMock:EmployeeProcedureRequest = {
       schema: "LEARCredentialEmployee",
       format: "jwt_vc_json",
       payload: {