+++ src/app/features/credential-details/credential-details.routes.ts
@@ -1,0 +1,11 @@
+import { Routes } from '@angular/router';
+import { CredentialDetailsComponent } from './credential-details.component';
+
+const routes: Routes = [
+  {
+    path: ':id',
+    component: CredentialDetailsComponent
+  }
+];
+
+export default routes;