+++ src/app/shared/components/form-credential/form-credential.component.ts
@@ -1,4 +1,4 @@
-import { Component, DestroyRef, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild, inject} from '@angular/core';
+import { Component, DestroyRef, Input, OnDestroy, OnInit, ViewChild, inject} from '@angular/core';
 import { FormGroupDirective, FormsModule } from '@angular/forms';
 import { CredentialProcedureService } from 'src/app/core/services/credential-procedure.service';
 import { Country, CountryService } from './services/country.service';
@@ -6,10 +6,10 @@ import { FormCredentialService } from './services/form-credential.service';
 import { AuthService } from 'src/app/core/services/auth.service';
 import { Router, RouterLink } from '@angular/router';
 import { MatOption } from '@angular/material/core';
-import { TempPower } from "../../../core/models/temporal/temp-power.interface";
+import { TempPower } from "../../../core/models/temp/temp-power.interface";
 import { TranslateService, TranslatePipe } from '@ngx-translate/core';
 import { from, Observable, of, switchMap, tap } from 'rxjs';
-import { Mandatee, OrganizationDetails, Power } from 'src/app/core/models/entity/lear-credential-employee.entity';
+import { CommonIssuer, EmployeeMandatee, StrictPower } from 'src/app/core/models/entity/lear-credential';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { MatButton } from '@angular/material/button';
 import { PowerComponent } from '../power/power.component';
@@ -63,17 +63,13 @@ import { MatIcon } from '@angular/material/icon';
 })
 export class FormCredentialComponent implements OnInit, OnDestroy {
   @ViewChild('formDirective') public formDirective!: FormGroupDirective;
-  @Output() public sendReminder = new EventEmitter<void>();
-  @Output() public signCredential = new EventEmitter<void>();
-  @Input({ required: true }) public viewMode: 'create' | 'detail' = 'create';
   @Input() public asSigner: boolean = false;
-  @Input() public isDisabled: boolean = false;
   @Input() public title: string = '';
-  @Input() public power: Power[] = [];
+  @Input() public power: StrictPower[] = [];
   @Input() public credentialStatus: string = '';
-  @Input() public issuer: OrganizationDetails = this.initializeOrganizationDetails();
-  @Input() public credential: Mandatee = this.initializeCredential();
-  @Input() public mandator: OrganizationDetails = this.initializeOrganizationDetails();
+  @Input() public issuer: CommonIssuer = this.initializeOrganizationDetails();
+  @Input() public credential: EmployeeMandatee = this.initializeCredential();
+  @Input() public mandator: CommonIssuer = this.initializeOrganizationDetails();
   public isLoading$: Observable<boolean>;
 
   public addedPowers$: Observable<TempPower[]>;
@@ -105,7 +101,7 @@ export class FormCredentialComponent implements OnInit, OnDestroy {
       .pipe(takeUntilDestroyed(this.destroyRef))
       .subscribe(mandator2 => {
         if (mandator2) {
-          if (this.viewMode === "create" && !this.asSigner) {
+          if (!this.asSigner) {
             this.mandator = {
               organizationIdentifier: mandator2.organizationIdentifier,
               organization: mandator2.organization,
@@ -118,9 +114,6 @@ export class FormCredentialComponent implements OnInit, OnDestroy {
         }
       });
 
-    if (this.viewMode === 'detail') {
-      this.tempPowers = this.power.map(power => this.formService.convertToTempPower(power));
-    }
   }
 
   public openSubmitDialog() {
@@ -182,14 +175,6 @@ export class FormCredentialComponent implements OnInit, OnDestroy {
     return this.router.navigate(['/organization/credentials']);
   }
 
-  public triggerSendReminder(): void {
-    this.sendReminder.emit();
-  }
-
-  public triggerSignCredential(): void {
-    this.signCredential.emit();
-  }
-
   public hasSelectedPower(): boolean {
     return this.formService.hasSelectedPower();
   }
@@ -198,20 +183,11 @@ export class FormCredentialComponent implements OnInit, OnDestroy {
     return this.formService.powersHaveFunction();
   }
 
-  public showReminderButton(): boolean {
-    return (this.viewMode === 'detail') && ((this.credentialStatus === 'WITHDRAWN') || (this.credentialStatus === 'DRAFT') || (this.credentialStatus === 'PEND_DOWNLOAD'));
-  }
-
-  public showSignCredentialButton(): boolean{
-    return (this.viewMode === 'detail') && (this.credentialStatus === 'PEND_SIGNATURE')
-  }
-
-
   public ngOnDestroy(): void {
     this.formService.reset();
   }
 
-  private initializeCredential(): Mandatee {
+  private initializeCredential(): EmployeeMandatee {
     return {
       firstName: '',
       lastName: '',
@@ -220,7 +196,7 @@ export class FormCredentialComponent implements OnInit, OnDestroy {
     };
   }
 
-  private initializeOrganizationDetails(): OrganizationDetails {
+  private initializeOrganizationDetails(): CommonIssuer {
     return {
       organizationIdentifier: '',
       organization: '',